refactor: move CountryPicker from bottomSheet to screen package
